provide support to E2E testing trough converting every unit tests that use fake external resources to integration test that runs with external resource that is used in production environment.
The tradeoff with this that debugging test using real implementation might be more challenging,
as not everything runs in the same process,
and also the speed of unit tests heavily affected by the external resource throughput.
Also since real external resource is a singleton on high level, parallel execution of the tests no longer available with E2E test execution mode.

The E2E testing mode is meant to be used in CI/CD pipelines,
and not for continous local development as it increase feedback loop.