merge deployment and release domains into one
The deployment package fundamentally was there only for the Environment entity,
and for some possible future features.

Since Environment is much closer to the SRP role actors of the release domain
and Environment has a high cohesion with Pilot and Rollout entities,
it makes sense to move it closer to them.