fix release rollout http api controllers context dependency
Because the context was not passed down to the httspec test,
when the test execution used Postgres with Tx as isolation,
the controller had no visibilty on the created example entities in the storage.

This is not an issue with in-memory fake testing double storage,
because they are created per test instead of isolated with a transaction.